Refactor `createChainableApi` into module, change `tabpage.window` to proxy window requests
