+++ src/api/Tabpage.js
@@ -1,4 +1,5 @@
 const BaseApi = require('./Base');
+const createChainableApi = require('./helpers/createChainableApi');
 
 class Tabpage extends BaseApi {
   // @return Promise<Array<Window>>
@@ -8,7 +9,12 @@ class Tabpage extends BaseApi {
 
   // @return Promise<Window>
   get window() {
-    return this.request(`${this.prefix}get_win`, [this]);
+    // Require is here otherwise we get circular refs
+    // eslint-disable-next-line global-require
+    const Window = require('./Window');
+    return createChainableApi.call(this, 'Window', Window, () =>
+      this.request(`${this.prefix}get_win`, [this])
+    );
   }
 
   // Is current tabpage valid
