+++ src/api/helpers/types.js
@@ -1,0 +1,13 @@
+const Buffer = require('../Buffer');
+const Window = require('../Window');
+const Tabpage = require('../Tabpage');
+
+const TYPES = {
+  Buffer,
+  Window,
+  Tabpage,
+};
+
+module.exports = TYPES;
+module.exports.default = module.exports;
+
